Separating Order Validation concerns
